Default TCP Congestion Control Algorithm Selection
Make all of the available TCP Congestion Control algorithms selectable
as the default. Also put all of them in alphabetical order so they are
easier to find and it is more obvious if any are missing.

Update the string table for suitable net.ipv4.tcp_congestion_control
values for missing algorithms, as reported by
net.ipv4.tcp_available_congestion_control.